chat-service/requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
python-jose[cryptography]